Repackaged classes for organization and added javadocs to a few.
