+++ build.gradle
@@ -18,6 +18,7 @@ dependencies {
     implementation 'org.reactivestreams:reactive-streams:1.0.4'
     implementation 'org.mongodb:mongodb-driver-reactivestreams:4.9.0'
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
+    testImplementation 'org.junit.jupiter:junit-jupiter:5.8.1'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
 }
