+++ src/main/java/org/bcit/com2522/project/scuffed/menu/Menu.java
@@ -1,4 +1,4 @@
-package org.bcit.com2522.project.scuffed.ui;
+package org.bcit.com2522.project.scuffed.menu;
 
 import org.bcit.com2522.project.scuffed.client.Window;
