+++ src/main/java/org/bcit/com2522/project/scuffed/menu/MenuState.java
@@ -1,9 +1,12 @@
-package org.bcit.com2522.project.scuffed.ui;
+package org.bcit.com2522.project.scuffed.menu;
 
 import org.bcit.com2522.project.scuffed.client.Window;
+import org.bcit.com2522.project.scuffed.uiComponents.Button;
+import org.bcit.com2522.project.scuffed.uiComponents.ButtonManager;
 
 /**
- * The type Menu state.
+ * The type Menu state. This class is the base class for all menu states. It contains a button manager a reference to the scene and menu.
+ * TODO: Remove the scene reference from this and all subclasses.
  */
 public abstract class MenuState {
   /**
