+++ src/main/java/org/bcit/com2522/project/scuffed/uiComponents/Clickable.java
@@ -1,4 +1,4 @@
-package org.bcit.com2522.project.scuffed.client;
+package org.bcit.com2522.project.scuffed.uiComponents;
 
 /**
  * The type Clickable.
