+++ src/main/java/org/bcit/com2522/project/scuffed/uiComponents/InputBox.java
@@ -1,4 +1,4 @@
-package org.bcit.com2522.project.scuffed.ui;
+package org.bcit.com2522.project.scuffed.uiComponents;
 
 import org.bcit.com2522.project.scuffed.client.Window;
 import processing.core.PApplet;
